The diff includes changes to the CSS file, controller file, header file, and routes file.
In the CSS file, a new class called `.btn-hello` is created. It has the properties of duration, delay, and animation. The `.btn-green` class is also updated to have a white background and a gray border.

In the controller file, the `configure_permitted_parameters` method is updated to allow for the parameters `:first_name`, `:last_name`, and `:user_name` to be permitted for both sign up and account update.

In the header file, the navigation bar is updated to include a new link to the `new_book_path`. The link is only visible to logged-in users.

In the routes file, the `root` path is updated to point to the `books#index` action.