Add-Devise-authentication,-Users-model,-Add-tests (#2)
* ✨ feat:

add device gem

* ✨ feat:

add user model

* ✨ feat:

generate devise:views

* ✨ feat:

add validates for  first_name, last_name

* 🐞 fix:

correct validation

* 🌈 style:

user log

* 🌈 style:

add

* 🧪 test:

test

* 🦄 refactor:

del unnecessary view

* 🧪 test:

user_spec test

* 🐞 fix:

fix warning from expert

* ✨ feat:

add gem Inline SVG and change _header

* The diff includes changes to the CSS file, controller file, header file, and routes file.

In the CSS file, a new class called `.btn-hello` is created. It has the properties of duration, delay, and animation. The `.btn-green` class is also updated to have a white background and a gray border.

In the controller file, the `configure_permitted_parameters` method is updated to allow for the parameters `:first_name`, `:last_name`, and `:user_name` to be permitted for both sign up and account update.

In the header file, the navigation bar is updated to include a new link to the `new_book_path`. The link is only visible to logged-in users.

In the routes file, the `root` path is updated to point to the `books#index` action.

* Update routes.rb, user_spec.rb, and users_spec.rb

* Update styles for books index and show pages

* Update styles for books index, forms, devise registrations and sessions, and book factory

* Update the background color of the login and registration pages from #f0f0f0 to #707070

* removes devise_parameter_sanitizer.permit(:account_update, keys: [:first_name, :last_name, :user_...

* ```
This commit changes the primary abstract class of the ApplicationRecord model from ActiveRecord::Base to nil.

* Update books migration and add users migration

* Update the book show page and header

* The navigation bar has been simplified. The "Log Out" button has been moved to the right of the navigation bar. The "Sign in" and "Sign Up" buttons have been moved to the left of the navigation bar.

* Change tab

* Update the book show page layout

* The diff shows that the root path was changed from pointing to the `books#index` action to the `rails/health#show` action.

The `devise_for` line was also modified to only include the `:sessions` and `:registrations` actions.

* The diff removes the devise_for :users, only: [:sessions, :registrations] line, which was used to enable user authentication.

* 🌈 style:

Removed 3 instances of unnecessary whitespace.

* 🌈 style:

 Update application.html.erb

* 🦄 refactor:

 Updates conditional statement to check if current_user is present instead of truthy